Se arreglan errores en el controller de Donation y se agrega un nuevo Dto , para la respuesta del servicio con las entiedas de User , referenciando el ID del mismio

+++ surotec-academy/src/main/java/com/website/surotec_academy/domain/dto/donation/DonationCreateDto.java
@@ -2,6 +2,6 @@ package com.website.surotec_academy.domain.dto.donation;
 
 public record DonationCreateDto(
         String description,
-        Double amount,
+        Integer amount,
         Long idUser
 ) {}
